Address PR review feedback
- Fix unused imports/variables (Legend, COLORS, res) for strict tsc -b
- Fix Recharts Tooltip formatter types (ValueType compatibility)
- Move import type to top of client.ts (verbatimModuleSyntax)
- Make API base URL configurable via VITE_API_BASE_URL env var
- Add error handling to PlantConfigPage save()
- Fix cash flow table year numbering (1-based to match library)
- Fix additional_capex_years schema type: list[float] -> list[int]
- Remove invalid ${PROJECT_ROOT} dep from backend pyproject.toml
- Add pip install -e . step to README/docs run instructions
- Fix React 18 -> 19 in GUI_ARCHITECTURE.md

+++ backend/app/schemas.py
@@ -58,7 +58,7 @@ class PlantConfigIn(BaseModel):
     plant_products: dict[str, dict] = Field(default_factory=dict)
     fc: float | None = None
     fp: float | None = None
-    additional_capex_years: list[float] | None = None
+    additional_capex_years: list[int] | None = None
     additional_capex_cost: list[float] | None = None
